refactor(scan-nh): move scoring code to its own module
